Temporarily disable pkgconf whenever building the kernel.
I found out that I can't just "emerge -C pkgconf" then bring it
back after running `genkernel`. That's because genkernel-next
depends on `pkg-config` existing (which is also what makes my
kernel fail to compile, currently). The only way I see to resolve
this right now is to go into genkernel's shell scripts and tell
it where it needs to turn pkgconf "on" and "off" so that both
of these things can succeed.